chore(config): deactivate recurring job for past appointment cancellation
The `CancelPastAppointmentsJob` has been disabled in all environments by commenting it out in the recurring jobs configuration.

This temporary measure is required to support a new workflow that allows for the editing of past appointments. The automatic cancellation was preventing this by changing the appointment state. The job can be re-evaluated and potentially re-enabled in the future.